+++ examples/data-sources/fmc_cipher_suite_lists/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_cipher_suite_lists" "example" {
+  items = {
+    "my_cipher_suite_lists" = {
+    }
+  }
+}