+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- (Enhancement) Add support for `fmc_cipher_suite_list` and `fmc_cipher_suite_lists`
 - (Enhancement) Add support for `fmc_mac_address_pool` and `fmc_mac_address_pools`
 - (Fix) `fmc_chassis_physical_interface` is not cleared correctly on destroy
 - (Fix) `fmc_device_ha_pair` may not detect unsuccessful split request